+++ cluster/src/main/scala/io/snappydata/ToolsCallbackImpl.scala
@@ -22,16 +22,12 @@ import java.net.URLClassLoader
 
 import com.pivotal.gemfirexd.internal.engine.Misc
 import com.pivotal.gemfirexd.internal.engine.distributed.utils.GemFireXDUtils
-import com.pivotal.gemfirexd.internal.iapi.error.StandardException
-import com.pivotal.gemfirexd.internal.iapi.util.IdUtil
-import com.pivotal.gemfirexd.internal.impl.jdbc.{EmbedConnection, Util}
-import com.pivotal.gemfirexd.internal.impl.sql.execute.PrivilegeInfo
-import com.pivotal.gemfirexd.internal.shared.common.reference.SQLState
 import io.snappydata.cluster.ExecutorInitiator
 import io.snappydata.impl.LeadImpl
 
 import org.apache.spark.executor.SnappyExecutor
 import org.apache.spark.sql.execution.columnar.ExternalStoreUtils
+import org.apache.spark.sql.execution.columnar.impl.StoreCallbacksImpl
 import org.apache.spark.sql.execution.ui.SQLTab
 import org.apache.spark.sql.hive.thriftserver.SnappyHiveThriftServer2
 import org.apache.spark.ui.{JettyUtils, SnappyDashboardTab}
@@ -164,38 +160,6 @@ object ToolsCallbackImpl extends ToolsCallback with Logging {
     ret
   }
 
-  override def checkSchemaPermission(schema: String, currentUser: String): String = {
-    val ms = Misc.getMemStoreBootingNoThrow
-    val userId = IdUtil.getUserAuthorizationId(currentUser)
-    if (ms ne null) {
-      var conn: EmbedConnection = null
-      if (ms.isSnappyStore && Misc.isSecurityEnabled) {
-        var contextSet = false
-        try {
-          val dd = ms.getDatabase.getDataDictionary
-          conn = GemFireXDUtils.getTSSConnection(false, true, false)
-          conn.getTR.setupContextStack()
-          contextSet = true
-          val sd = dd.getSchemaDescriptor(
-            schema, conn.getLanguageConnection.getTransactionExecute, false)
-          if (sd eq null) {
-            if (schema.equalsIgnoreCase(userId) ||
-                schema.equalsIgnoreCase(userId.replace('-', '_'))) {
-              if (ms.tableCreationAllowed()) return userId
-              throw StandardException.newException(SQLState.AUTH_NO_ACCESS_NOT_OWNER,
-                schema, schema)
-            } else {
-              throw StandardException.newException(SQLState.LANG_SCHEMA_DOES_NOT_EXIST, schema)
-            }
-          }
-          PrivilegeInfo.checkOwnership(userId, sd, sd, dd)
-          sd.getAuthorizationId
-        } catch {
-          case se: StandardException => throw Util.generateCsSQLException(se)
-        } finally {
-          if (contextSet) conn.getTR.restoreContextStack()
-        }
-      } else userId
-    } else userId
-  }
+  override def checkSchemaPermission(schema: String, currentUser: String): String =
+    StoreCallbacksImpl.checkSchemaPermission(schema, currentUser)
 }
